draft 3

+++ Domain/Combinations/BenchmarkProgress.cs
@@ -1,10 +1,10 @@
 ﻿namespace Domain.Combinations
 {
-	public class BenchmarkProgress
-	{
-		public int CurrentIteration { get; set; }
-		public int TotalIterations { get; set; }
-		public int PercentComplete { get; set; }
-		public string StatusMessage { get; set; }
-	}
+    public class BenchmarkProgress
+    {
+        public ulong CurrentIteration { get; set; }
+        public ulong TotalIterations { get; set; }
+        public double PercentComplete { get; set; }
+        public string StatusMessage { get; set; }
+    }
 }
